Added a webpage flow, deleted unnecessary files
