+++ frontend/src/App.jsx
@@ -6,23 +6,20 @@ import {
 } from "react-router-dom";
 import SignUp from "./pages/SignUp";
 import SignIn from "./pages/SignIn";
-import SignUpForm from "./pages/SignUpForm";
 import PhotoGallery from "./pages/PhotoGallery";
-import HomePage from "./pages/HomePage";
 import Onboarding from "./pages/Onboarding";
 import UserPage from "./pages/user";
 import Restaurant from "./pages/Restaurants";
 import Review from "./pages/Review";
 import Header from "./components/header";
-import MapSandbox from "./pages/mapSandbox";
 import AccountSettings from "./pages/AccountSettings";
 import ReviewPage from "./pages/ReviewPage";
-
 function AppLayout() {
    const location = useLocation();
    const hideHeaderPaths = [
       "/",
       "/signin",
+      "/signup",
       "/signup-form",
       "/onboarding",
    ];
@@ -45,13 +42,9 @@ function AppLayout() {
             style={{ flex: 1, overflow: "auto" }}
          >
             <Routes>
-               <Route path="/" element={<SignUp />} />
+               <Route path="/" element={<SignIn />} />
                <Route path="/signin" element={<SignIn />} />
-               <Route
-                  path="/signup-form"
-                  element={<SignUpForm />}
-               />
-               <Route path="/home" element={<HomePage />} />
+               <Route path="/signup" element={<SignUp />} />
                <Route
                   path="/onboarding"
                   element={<Onboarding />}
@@ -66,10 +59,6 @@ function AppLayout() {
                   element={<Restaurant />}
                />
                <Route path="/review" element={<Review />} />
-               <Route
-                  path="/sandbox"
-                  element={<MapSandbox />}
-               />
                <Route
                   path="/settings"
                   element={<AccountSettings />}
