+++ frontend/src/pages/SignIn.jsx
@@ -1,17 +1,17 @@
 import React, { useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import "./auth.css";
-import logo from "../assets/logo.png";
 import bgImage from "../assets/signup-bg.jpeg";
+
 export default function SignIn() {
    const navigate = useNavigate();
 
    const [email, setEmail] = useState("");
    const [password, setPassword] = useState("");
 
-   const handleSignIn = (e) => {
+   const handleSignIn = async (e) => {
       e.preventDefault();
-      navigate("/onboarding");
+      navigate("/restaurants");
    };
 
    return (
@@ -26,13 +26,7 @@ export default function SignIn() {
             role="dialog"
             aria-label="Sign in"
          >
-            <div className="auth__brand">
-               <img
-                  src={logo}
-                  alt="Umami logo"
-                  className="auth__logo"
-               />
-            </div>
+            <div className="auth__brand">umami</div>
 
             <h1 className="auth__title">
                Sign into your account
@@ -53,7 +47,6 @@ export default function SignIn() {
                   value={email}
                   onChange={(e) => setEmail(e.target.value)}
                   required
-                  aria-label="Email"
                />
 
                <input
@@ -66,7 +59,6 @@ export default function SignIn() {
                      setPassword(e.target.value)
                   }
                   required
-                  aria-label="Password"
                />
 
                <button
@@ -80,12 +72,13 @@ export default function SignIn() {
                   <span>Don't have an account?</span>
                </div>
 
-               <Link
-                  to="/signup-form"
+               <button
                   className="auth__secondary"
+                  type="button"
+                  onClick={() => navigate("/signup")}
                >
                   Sign Up
-               </Link>
+               </button>
 
                <p className="auth__legal">
                   By continuing, you agree to our{" "}
