--- frontend/src/pages/home.jsx
@@ -1,11 +1,0 @@
-// display hello world
-
-function Home() {
-   return (
-      <div>
-         <h1>Hello, World!</h1>
-      </div>
-   );
-}
-
-export default Home;
