--- frontend/src/pages/homepage.css
@@ -1,36 +1,0 @@
-.homepage {
-   min-height: 100vh;
-   background-color: #f4f2ea;
-   font-family: "Poppins", sans-serif;
-}
-
-.home-hero {
-   text-align: center;
-   margin-top: 120px;
-   padding: 0 20px;
-}
-
-.home-hero h2 {
-   font-size: 40px;
-   margin-bottom: 20px;
-}
-
-.home-hero p {
-   font-size: 18px;
-   color: #555;
-   margin-bottom: 30px;
-}
-
-.explore-btn {
-   padding: 12px 24px;
-   font-size: 16px;
-   border: none;
-   background-color: #2f5d3a;
-   color: white;
-   border-radius: 6px;
-   cursor: pointer;
-}
-
-.explore-btn:hover {
-   background-color: #244a2f;
-}
